fix: we use git rebase
