+++ src/test/java/selenide/RestaurantReservationTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import selenide.pages.*;
 
 import static com.codeborne.selenide.Selenide.open;
-import static selenide.pages.Page.BASE_URL;
+import static selenide.pages.Page.*;
 
 public class RestaurantReservationTest {
 
@@ -26,8 +26,8 @@ public class RestaurantReservationTest {
         loginPage = homepage.clickOnLoginBtn();
         loginPage.fillInValidCred();
         loginPage.clickOnLoginBtnHP();
-        filterPage = homepage.fillInTheRestaurantAndClickOnFindBtn();
-        restaurantPage = filterPage.clickOnRestaurant();
+        filterPage = homepage.searchRestaurant(SAMPLE_RESTAURANT,SAMPLE_CITY);
+        restaurantPage = filterPage.clickOnFirstRestaurantBtn();
         completeReservationPage = restaurantPage.reserveNow("7:00 pm", "7:30 pm");
         checkoutSummaryPage = completeReservationPage.clickOnReserveNowBtn();
         userReservationsPage = checkoutSummaryPage.clickOnLoginIcon();
